[VideoDB] Move DDL functions and dependencies into VideoDatabaseDDL.cpp
